+++ hoot-core/src/main/cpp/hoot/core/visitors/CountUniqueReviewsVisitor.h
@@ -22,7 +22,7 @@
  * This will properly maintain the copyright information. Maxar
  * copyrights will be updated automatically.
  *
- * @copyright Copyright (C) 2015, 2016, 2017, 2018, 2019, 2020, 2021 Maxar (http://www.maxar.com/)
+ * @copyright Copyright (C) 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022 Maxar (http://www.maxar.com/)
  */
 #ifndef COUNTUNIQUEREVIEWSVISITOR_H
 #define COUNTUNIQUEREVIEWSVISITOR_H
@@ -48,7 +48,7 @@ public:
   CountUniqueReviewsVisitor() = default;
   ~CountUniqueReviewsVisitor() override = default;
 
-  double getStat() const override { return _reviews.size(); }
+  double getStat() const override { return static_cast<double>(_reviews.size()); }
 
   /**
    * @see ElementVisitor
